Improve MessageBoxes helper
The helper to create delete confirmation messages only offers limited
functionality. For customised messages such as setting the error level
or tuning the header text, the helpers logic has to be duplicated.

To solve this, the creation of the messagebox is splitt into a separate
function.
The show function still exists to ensure backwards compatibility.
Also the item list has been tuned to show how many items are not shown
in the conformation message.

288246