RemoteFileService: enhance locale handling
The remote file service uses the given locale to find language dependant
templates. This logic is too complex and can be simplified. Also, the
implemented security pattern is too restricting (i.e. not allowing # in
locales even tho # is valid locale character). Last but not least, there
is little comments on the complicated code parts and no unit test at all.

The security pattern is changed into a simple allow list of expected
locale characters.
The part which checks if there is already a matching locale at the end
of the filename is simplified.
Test caseses for the new function are added.

382967